# api/openapi.yaml
openapi: 3.0.3
info:
  title: Music Tools API
  version: 0.1.0
paths:
  /api/v1/scales:
    get:
      summary: List all scales
      description: Returns the list of available scale definitions.
      parameters:
        - name: name
          in: query
          description: Filter by scale name or common name (case-insensitive).
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Scale definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScaleListResponse"
  /api/v1/scales/{id}:
    get:
      summary: Get a scale by id
      description: Returns a single scale definition by id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scale definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScaleDefinition"
        "404":
          description: Scale not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/scales/random:
    get:
      summary: Get a random scale
      description: Returns a random diatonic scale selection for the given scale filters and accidental limits.
      parameters:
        - name: maxAccidentals
          in: query
          description: Maximum number of sharps or flats allowed (0-7).
          required: false
          schema:
            type: integer
            default: 5
            minimum: 0
            maximum: 7
        - name: scale
          in: query
          description: Scale name or common name. Repeat to provide multiple scale filters.
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Random scale selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RandomScaleResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/scales/scale_layouts:
    get:
      summary: List scale layouts by tuning
      description: Returns scale layouts grouped by tuning and scale.
      responses:
        "200":
          description: Scale layouts by tuning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutInstanceListResponse"
  /api/v1/scales/scale_layouts/{tuningId}:
    get:
      summary: Get scale layouts by tuning id
      description: Returns scale layouts for a specific tuning id.
      parameters:
        - name: tuningId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scale layouts for tuning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutTuningInstance"
        "404":
          description: Layout tuning not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tunings:
    get:
      summary: List tunings
      description: Returns all tuning definitions.
      parameters:
        - name: name
          in: query
          description: Filter by tuning name (exact match).
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Tuning definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TuningListResponse"
  /api/v1/tunings/{id}:
    get:
      summary: Get tuning by id
      description: Returns a tuning definition by id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tuning definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TuningDefinition"
        "404":
          description: Tuning not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ScaleDefinition:
      type: object
      required:
        - id
        - name
        - common_name
        - type
        - intervals
      properties:
        id:
          type: integer
          description: Stable scale identifier.
        name:
          type: string
          description: Scale name.
        common_name:
          type: string
          description: Common scale name.
        type:
          type: string
          description: Scale category.
        intervals:
          type: array
          items:
            type: integer
          description: Semitone intervals from the root.
    ScaleListResponse:
      type: object
      required:
        - scales
      properties:
        scales:
          type: array
          items:
            $ref: "#/components/schemas/ScaleDefinition"
    RandomScaleResponse:
      type: object
      required:
        - scale
        - key
        - notes
      properties:
        scale:
          type: string
          description: Scale name.
        key:
          type: string
          description: Root key.
        notes:
          type: array
          items:
            type: string
          description: Note names in the selected scale.
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    FretRange:
      type: object
      required:
        - start
        - span
      properties:
        start:
          type: integer
        span:
          type: integer
    TuningDefinition:
      type: object
      required:
        - id
        - name
        - string_count
        - strings
      properties:
        id:
          type: integer
        name:
          type: string
        string_count:
          type: integer
        strings:
          type: array
          items:
            type: string
    TuningListResponse:
      type: object
      required:
        - tunings
      properties:
        tunings:
          type: array
          items:
            $ref: "#/components/schemas/TuningDefinition"
    LayoutInstanceListResponse:
      type: object
      required:
        - tunings
      properties:
        tunings:
          type: array
          items:
            $ref: "#/components/schemas/LayoutTuningInstance"
    LayoutTuningInstance:
      type: object
      required:
        - id
        - name
        - string_count
        - strings
        - scales
      properties:
        id:
          type: integer
        name:
          type: string
        string_count:
          type: integer
        strings:
          type: array
          items:
            type: string
        scales:
          type: array
          items:
            $ref: "#/components/schemas/LayoutScaleInstance"
    LayoutScaleInstance:
      type: object
      required:
        - id
        - name
        - type
        - positions
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        positions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/LayoutPositionInstance"
    LayoutPositionInstance:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
        start:
          type: integer
        span:
          type: integer
        per_string:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FretRange"
        per_string_frets:
          type: object
          additionalProperties:
            type: array
            items:
              type: integer
        validated_manual:
          type: boolean